+++ packages/manager/src/features/Backups/AutoEnroll.tsx
@@ -1,12 +1,12 @@
 import { styled } from '@mui/material/styles';
 import * as React from 'react';
 
+import { FormControlLabel } from 'src/components/FormControlLabel';
 import { Link } from 'src/components/Link';
 import { Notice } from 'src/components/Notice/Notice';
 import { Paper } from 'src/components/Paper';
 import { Toggle } from 'src/components/Toggle';
 import { Typography } from 'src/components/Typography';
-import FormControlLabel from 'src/components/core/FormControlLabel';
 
 interface AutoEnrollProps {
   enabled: boolean;