+++ packages/manager/src/features/Firewalls/FirewallDetail/Rules/FirewallRuleForm.tsx
@@ -8,8 +8,8 @@ import { Notice } from 'src/components/Notice/Notice';
 import { Radio } from 'src/components/Radio/Radio';
 import { TextField } from 'src/components/TextField';
 import { Typography } from 'src/components/Typography';
-import FormControlLabel from 'src/components/core/FormControlLabel';
-import RadioGroup from 'src/components/core/RadioGroup';
+import { FormControlLabel } from 'src/components/FormControlLabel';
+import { RadioGroup } from 'src/components/RadioGroup';
 import {
   addressOptions,
   firewallOptionItemsShort,