+++ packages/manager/src/features/Support/SupportTickets/SupportTicketDialog.tsx
@@ -5,7 +5,6 @@ import {
 import { APIError } from '@linode/api-v4/lib/types';
 import { Theme } from '@mui/material/styles';
 import { makeStyles } from '@mui/styles';
-import * as Bluebird from 'bluebird';
 import { update } from 'ramda';
 import * as React from 'react';
 import { debounce } from 'throttle-debounce';
@@ -14,11 +13,12 @@ import { Accordion } from 'src/components/Accordion';
 import { ActionsPanel } from 'src/components/ActionsPanel/ActionsPanel';
 import { Dialog } from 'src/components/Dialog/Dialog';
 import Select, { Item } from 'src/components/EnhancedSelect/Select';
+import { FormHelperText } from 'src/components/FormHelperText';
+import { Link } from 'src/components/Link';
 import { Notice } from 'src/components/Notice/Notice';
 import { EntityForTicketDetails } from 'src/components/SupportLink/SupportLink';
 import { TextField } from 'src/components/TextField';
 import { Typography } from 'src/components/Typography';
-import FormHelperText from 'src/components/core/FormHelperText';
 import { useAccount } from 'src/queries/account';
 import { useAllDatabasesQuery } from 'src/queries/databases';
 import { useAllDomainsQuery } from 'src/queries/domains';
@@ -32,6 +32,7 @@ import {
   getErrorMap,
   getErrorStringOrDefault,
 } from 'src/utilities/errorUtils';
+import { reduceAsync } from 'src/utilities/reduceAsync';
 import scrollErrorIntoView from 'src/utilities/scrollErrorIntoView';
 import { storage } from 'src/utilities/storage';
 
@@ -45,7 +46,6 @@ import SupportTicketSMTPFields, {
   smtpDialogTitle,
   smtpHelperText,
 } from './SupportTicketSMTPFields';
-import { Link } from 'src/components/Link';
 
 const useStyles = makeStyles((theme: Theme) => ({
   expPanelSummary: {
@@ -352,7 +352,7 @@ export const SupportTicketDialog = (props: SupportTicketDialogProps) => {
     setFiles(newFiles);
   };
 
-  /* Reducer passed into Bluebird.reduce() below.
+  /* Reducer passed into reduceAsync (previously Bluebird.reduce) below.
    * Unfortunately, this reducer has side effects. Uploads each file and accumulates a list of
    * any upload errors. Also tracks loading state of each individual file. */
   const attachFileReducer = (
@@ -410,7 +410,7 @@ export const SupportTicketDialog = (props: SupportTicketDialogProps) => {
 
     /* Upload each file as an attachment, and return a Promise that will resolve to
      *  an array of aggregated errors that may have occurred for individual uploads. */
-    return Bluebird.reduce(filesWithTarget, attachFileReducer, {
+    return reduceAsync(filesWithTarget, attachFileReducer, {
       errors: [],
       success: [],
     });