+++ packages/manager/src/queries/linodes/configs.ts
@@ -1,10 +1,20 @@
 import {
   APIError,
   Config,
+  ConfigInterfaceOrderPayload,
+  Interface,
+  InterfacePayload,
   LinodeConfigCreationData,
+  UpdateConfigInterfacePayload,
+  appendConfigInterface,
   createLinodeConfig,
   deleteLinodeConfig,
+  deleteLinodeConfigInterface,
+  getConfigInterface,
+  getConfigInterfaces,
+  updateConfigInterface,
   updateLinodeConfig,
+  updateLinodeConfigOrder,
 } from '@linode/api-v4';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 
@@ -19,6 +29,10 @@ export const useAllLinodeConfigsQuery = (id: number, enabled = true) => {
   );
 };
 
+const configQueryKey = 'configs';
+const interfaceQueryKey = 'interfaces';
+
+// Config queries
 export const useLinodeConfigDeleteMutation = (
   linodeId: number,
   configId: number
@@ -32,7 +46,7 @@ export const useLinodeConfigDeleteMutation = (
           queryKey,
           'linode',
           linodeId,
-          'configs',
+          configQueryKey,
         ]);
       },
     }
@@ -49,7 +63,7 @@ export const useLinodeConfigCreateMutation = (linodeId: number) => {
           queryKey,
           'linode',
           linodeId,
-          'configs',
+          configQueryKey,
         ]);
       },
     }
@@ -69,7 +83,167 @@ export const useLinodeConfigUpdateMutation = (
           queryKey,
           'linode',
           linodeId,
-          'configs',
+          configQueryKey,
+        ]);
+      },
+    }
+  );
+};
+
+// Config Interface queries
+export const useConfigInterfacesQuery = (
+  linodeID: number,
+  configID: number
+) => {
+  return useQuery<Interface[], APIError[]>(
+    [
+      queryKey,
+      'linode',
+      linodeID,
+      configQueryKey,
+      'config',
+      configID,
+      interfaceQueryKey,
+    ],
+    () => getConfigInterfaces(linodeID, configID),
+    { keepPreviousData: true }
+  );
+};
+
+export const useConfigInterfaceQuery = (
+  linodeID: number,
+  configID: number,
+  interfaceID: number
+) => {
+  return useQuery<Interface, APIError[]>(
+    [
+      queryKey,
+      'linode',
+      linodeID,
+      configQueryKey,
+      'config',
+      configID,
+      interfaceQueryKey,
+      'interface',
+      interfaceID,
+    ],
+    () => getConfigInterface(linodeID, configID, interfaceID),
+    { keepPreviousData: true }
+  );
+};
+
+export const useConfigInterfacesOrderMutation = (
+  linodeID: number,
+  configID: number
+) => {
+  const queryClient = useQueryClient();
+  return useMutation<{}, APIError[], ConfigInterfaceOrderPayload>(
+    (data) => updateLinodeConfigOrder(linodeID, configID, data),
+    {
+      onSuccess() {
+        queryClient.invalidateQueries([
+          queryKey,
+          'linode',
+          linodeID,
+          configQueryKey,
+          'config',
+          interfaceQueryKey,
+        ]);
+      },
+    }
+  );
+};
+
+export const useAppendConfigInterfaceMutation = (
+  linodeID: number,
+  configID: number
+) => {
+  const queryClient = useQueryClient();
+  return useMutation<Interface, APIError[], InterfacePayload>(
+    (data) => appendConfigInterface(linodeID, configID, data),
+    {
+      onSuccess() {
+        queryClient.invalidateQueries([
+          queryKey,
+          'linode',
+          linodeID,
+          configQueryKey,
+          'config',
+          configID,
+          interfaceQueryKey,
+        ]);
+      },
+    }
+  );
+};
+
+export const useUpdateConfigInterfaceMutation = (
+  linodeID: number,
+  configID: number,
+  interfaceID: number
+) => {
+  const queryClient = useQueryClient();
+  return useMutation<Interface, APIError[], UpdateConfigInterfacePayload>(
+    (data) => updateConfigInterface(linodeID, configID, interfaceID, data),
+    {
+      onSuccess: (InterfaceObj) => {
+        queryClient.invalidateQueries([
+          queryKey,
+          'linode',
+          linodeID,
+          configQueryKey,
+          'config',
+          configID,
+          interfaceQueryKey,
+        ]);
+        queryClient.setQueryData<Interface>(
+          [
+            queryKey,
+            'linode',
+            linodeID,
+            configQueryKey,
+            'config',
+            configID,
+            interfaceQueryKey,
+            'interface',
+            InterfaceObj.id,
+          ],
+          InterfaceObj
+        );
+      },
+    }
+  );
+};
+
+export const useDeleteConfigInterfaceMutation = (
+  linodeID: number,
+  configID: number,
+  interfaceID: number
+) => {
+  const queryClient = useQueryClient();
+  return useMutation<{}, APIError[]>(
+    () => deleteLinodeConfigInterface(linodeID, configID, interfaceID),
+    {
+      onSuccess() {
+        queryClient.invalidateQueries([
+          queryKey,
+          'linode',
+          linodeID,
+          configQueryKey,
+          'config',
+          configID,
+          interfaceQueryKey,
+        ]);
+        queryClient.removeQueries([
+          queryKey,
+          'linode',
+          linodeID,
+          configQueryKey,
+          'config',
+          configID,
+          interfaceQueryKey,
+          'interface',
+          interfaceID,
         ]);
       },
     }