+++ CHANGELOG.md
@@ -3,6 +3,6 @@
 This changelog has been deprecated in favor individual changelogs for each package.
 
 See changelogs for:
-- [@linode/api-v4](https://github.com/linode/manager/packages/api-v4/CHANGELOG.md)
-- [@linode/manager](https://github.com/linode/manager/packages/manager/CHANGELOG.md)
-- [@linode/validation](https://github.com/linode/manager/packages/validation/CHANGELOG.md)
+- [@linode/api-v4](https://github.com/linode/manager/blob/master/packages/api-v4/CHANGELOG.md)
+- [@linode/manager](https://github.com/linode/manager/blob/master/packages/manager/CHANGELOG.md)
+- [@linode/validation](https://github.com/linode/manager/blob/master/packages/validation/CHANGELOG.md)